% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify}
\alias{classify}
\title{Function to classify gauge time-series (zoo object) based on some prescribed aggregation level.}
\usage{
classify(gauge.ts, aggregation_level = "season")
}
\arguments{
\item{gauge.ts}{A zoo object containing the gauge observations.}

\item{aggregation_level}{A character string defining the classification criteria.}
}
\value{
Object \code{ts.classification} A zoo object with a single data column defining the assigned class for each time-step.
}
\description{
This method classifies a gauge time-series based on a pre-defined classification ("aggregation_level"). Currently only a seasonal classification is implemented. The output is added to the gauge spatial object as an additional spatial data frame column.
}
\details{
If aggregation_level is "season", the date strings of the zoo object guage.ts are converted to seasonal indicators, i.e. 1 (DJF), 2 (MAM), 3 (JJA), 4 (SON).
}
\examples{
data(gauges) # STFDF object with gauge observations
gauge.sp <- gauges@sp
gauge.ts <- as(gauges[,,1],"xts")
colnames(gauge.ts) <- gauge.sp$estacion
ts.classification <- classify(gauge.ts,aggregation_level="season")
summary(ts.classification)
}

