% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurl_download.R
\name{rcurl_download}
\alias{rcurl_download}
\title{Function to download GPM-IMERG HDF5 files from NASA ftp server using RCurl}
\usage{
rcurl_download(product, nrt_type, start, end, userpwd)
}
\arguments{
\item{product}{A character string defining the GPM-IMERG product to be downloaded. This can be either "production" (gauged-corrected, 4-month latency) or "nrt" (Near-real time, without gauge correction, latency < 1 day).}

\item{nrt_type}{A character string defining the NRT product type to be downloaded. This can be either "early" (4-hr latency, raw satellite - no correction) or "late" (18-hr latency, satellite with climatological correction). The \code{nrt_type} is only relevant if \code{product} is set to "nrt" and is ignored otherwise.}

\item{start}{A numeric value defining the starting date (full day) for the download in the format: YYYYMMDD (e.g. 20150401).}

\item{end}{A numeric value defining the ending date (full day) for the download in the format: YYYYMMDD (e.g. 20150402).}

\item{userpwd}{A character string defining your username and password to access the NASA ftp server. Downloading IMERG HDF5 files from the NASA ftp servers requires you to register your email here http://registration.pps.eosdis.nasa.gov/registration/newContact.html. Ensure to check the box that you are also interested in NRT products, if this is the case. Once you have done this and received the confirmation email enter your username and password in this format ("username:password"): "you(at)email:you(at)email". Your email == your username == your password.}
}
\value{
download_list A data.frame obect with a single column containing the character strings that define the file names the downloaded HDF 5 files. Additionally, the HDF5 files between the specified start and end dates are written to a sub-directory of the local working directory, named ("IMERG_HDF").
}
\description{
This function downloads GPM-IMERG HDF5 files from the NASA PPS ftp servers and writes these to a sub-directory of the local working directory, named ("IMERG_HDF"). Unlike \code{system_download} it does nmot require a dpwnload_list to be provided externally, but creates one itself. Go to http://pmm.nasa.gov/data-access/downloads/gpm for more info on the products.
}
\details{
This function downloads GPM-IMERG HDF5 files from the start to the end dates from the NASA ftp servers using RCurl. DIFFERENCE BETWEEN \code{system_download} & \code{rcurl_download}: \code{system_download} differs from \code{rcurl_download} in that it uses tools (wget, curl) external to R to download the HDF5 files from the NASA Mirador http server, whereas \code{rcurl_download} utilizes RCurl to perform the download entirely from within R. \code{rcurl_download} also obtains the data from NASA ftp servers. This requires users to register their email and supply email and password to the \code{rcurl_download} but requires no further external steps and generates its own download list within specified dates, whereas \code{system_download} requires the download list to be generated externally and loaded into R. However, no registration is required.
}
\examples{
data(production_url_list) # list of GPM IMERG HDF5 files including NASA http URL
product <- "production"
nrt_type <- "late"
start <- 20150401
end <- 20150402
userpwd <- "you(at)email:you(at)email"
# Not Run rcurl_download(product,nrt_type,start,end,userpwd)
}

