% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/system_download.R
\name{system_download}
\alias{system_download}
\title{Function to download GPM-IMERG HDF5 files from NASA http server}
\usage{
system_download(method = "wget", download_list)
}
\arguments{
\item{method}{A character string defining which method is used for the batch download. This can be either "wget" or "curl". Users are responsible for ensuring their machine has the required functionality to run the selected method. Wget can be installed from here: http://gnuwin32.sourceforge.net/packages/wget.htm (for Windows). Ensure the executable is placed in the relevant directory.}

\item{download_list}{A data.frame with a single column containing character strings that define the URLs of the GPM-IMERG HDF5 files to be downloaded from the NASA Mirador http server. See data(production_url_list) for an example.}
}
\value{
No object is returned within R. However, the HDF5 files specified in the download list are written to a sub-directory of the local working directory, named ("IMERG_HDF").
}
\description{
This function downloads GPM-IMERG HDF5 files from the NASA PPS Mirador using a pre-specified list of URLs and writes these to a sub-directory of the local working directory, named ("IMERG_HDF"). Go to http://pmm.nasa.gov/data-access/downloads/gpm for more info on the products.
}
\details{
This function downloads GPM-IMERG HDF5 files with global coverage as specified in the download list. Details on how to generate the download list are provided in the material accompanying this R-package. Users should check the length of the download list as the HDF5 files are about 2MB in size each. DIFFERENCE BETWEEN \code{system_download} & \code{rcurl_download}: \code{system_download} differs from \code{rcurl_download} in that it uses tools (wget, curl) external to R to download the HDF5 files from the NASA Mirador http server, whereas \code{rcurl_download} utilizes RCurl to perform the download entirely from within R. \code{rcurl_download} also obtains the data from NASA ftp servers. This requires users to register their email and supply email and password to the \code{rcurl_download} but requires no further external steps and generates its own download list within specified dates, whereas \code{system_download} requires the download list to be generated externally and loaded into R. However, no registration is required.
}
\examples{
data(production_url_list) # list of GPM IMERG HDF5 files including NASA http URL
product <- "production"
method <- "wget"
# Not Run system_download(method,download_list=production_url_list)
}

